% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundestag16}
\alias{bundestag16}
\title{16th electoral term}
\format{A data frame with 88608 rows and 5 variables.
\describe{
\item{Party}{Party associated with voter}
\item{Name}{Last name of voter}
\item{Full Name}{Full name (plus constituency in case of duplicate names) of voter}
\item{Vote ID}{ID of the role call vote, corresponding to electoral term and date}
\item{Vote}{Vote - one of yes, no, invalid, abstained}
}}
\usage{
bundestag16
}
\description{
A dataset containing role call votes cast during the 16th electoral term of the German Bundestag (18-10-2005 to 27-10-2009).
}
\keyword{datasets}
