% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundestag18}
\alias{bundestag18}
\title{18th electoral term}
\format{A data frame with 852 rows and 107 variables.
\describe{
\item{Fraktion/Gruppe}{Party associated with voter}
\item{Name}{Last name of voter}
\item{Bezeichnung}{Full name of voter}
...
}}
\usage{
bundestag18
}
\description{
A dataset containing role call votes cast during the 18th electoral term of the German Bundestag (22-10-2013 to 24-10-2017).
}
\keyword{datasets}
