% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundestag19}
\alias{bundestag19}
\title{19th electoral term}
\format{A data frame with 852 rows and 107 variables.
\describe{
\item{Fraktion/Gruppe}{Party associated with voter}
\item{Name}{Last name of voter}
\item{Bezeichnung}{Full name of voter}
...
}}
\usage{
bundestag19
}
\description{
A dataset containing role call votes cast during the 19th electoral term of the German Bundestag (24-10-2017).
}
\keyword{datasets}
