% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundestag17}
\alias{bundestag17}
\title{17th electoral term}
\format{A data frame with 852 rows and 107 variables.
\describe{
\item{Fraktion/Gruppe}{Party associated with voter}
\item{Name}{Last name of voter}
\item{Bezeichnung}{Full name of voter}
...
}}
\usage{
bundestag17
}
\description{
A dataset containing role call votes cast during the 17th electoral term of the German Bundestag (27-10-2009 to 22-10-2013).
}
\details{
All columns other than the ones described below are vote IDs and contain the following factors reflecting the voters' votes: `ja` - yes, `nein` - no, `ungültig` - invalid, `Enthaltung` - abstention, `nicht abgegeben` - vote not cast.
}
\keyword{datasets}
