% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundestag18}
\alias{bundestag18}
\title{18th electoral term}
\format{A data frame with 141404 rows and 5 variables.
\describe{
\item{Party}{Party associated with voter}
\item{Name}{Last name of voter}
\item{Full Name}{Full name (plus constituency in case of duplicate names) of voter}
\item{Vote ID}{ID of the role call vote, corresponding to electoral term and date}
\item{Vote}{Vote - one of yes, no, invalid, abstained}
}}
\usage{
bundestag18
}
\description{
A dataset containing role call votes cast during the 18th electoral term of the German Bundestag (22-10-2013 to 24-10-2017).
}
\keyword{datasets}
